PLAT-6803 - Moved service calls to the (created) invoker classes
