+++ integration-healthcheck/src/main/java/org/symphonyoss/integration/healthcheck/services/invokers/KmAuthHealthInvoker.java
@@ -1,4 +1,4 @@
-package org.symphonyoss.integration.healthcheck.services;
+package org.symphonyoss.integration.healthcheck.services.invokers;
 
 import static org.symphonyoss.integration.healthcheck.properties.HealthCheckProperties.IO_EXCEPTION;
 
@@ -6,29 +6,19 @@ import com.fasterxml.jackson.databind.JsonNode;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.actuate.health.Status;
-import org.springframework.context.annotation.Lazy;
-import org.springframework.stereotype.Component;
 import org.symphonyoss.integration.authentication.api.enums.ServiceName;
+import org.symphonyoss.integration.healthcheck.services.IntegrationBridgeServiceInfo;
 import org.symphonyoss.integration.json.JsonUtils;
 
 import java.io.IOException;
 
-/**
- * Service health indicator for Key Manager Authentication.
- *
- * Created by hamitay on 30/10/17.
- */
-@Component
-@Lazy
-public class KmAuthHealthIndicator extends AuthenticationServiceHealthIndicator {
+public class KmAuthHealthInvoker extends AuthenticationServiceHealthInvoker {
 
-  private static final Logger LOG = LoggerFactory.getLogger(KmAuthHealthIndicator.class);
-
+  private static final Logger LOG = LoggerFactory.getLogger(KmAuthHealthInvoker.class);
   private static final String SERVICE_FIELD = "keyauth";
-
   private static final String HC_AGGREGATED_URL_PATH = "/webcontroller/HealthCheck/aggregated";
-
   private static final String HC_URL_PATH = "/HealthCheck/aggregated";
+
 
   @Override
   protected String getHealthCheckUrl() {
@@ -36,7 +26,17 @@ public class KmAuthHealthIndicator extends AuthenticationServiceHealthIndicator 
   }
 
   @Override
-  protected void handleHealthResponse(IntegrationBridgeService service, String healthResponse) {
+  protected String getServiceBaseUrl() {
+    return properties.getKeyManagerAuthUrl();
+  }
+
+  @Override
+  protected String getServiceField() {
+    return SERVICE_FIELD;
+  }
+
+  @Override
+  protected void handleHealthResponse(IntegrationBridgeServiceInfo service, String healthResponse) {
     JsonNode jsonNode = parseJsonResponse(healthResponse);
 
     if (jsonNode == null) {
@@ -45,6 +45,25 @@ public class KmAuthHealthIndicator extends AuthenticationServiceHealthIndicator 
 
     Status status = retrieveConnectivityStatus(jsonNode);
     service.setConnectivity(status);
+  }
+
+  @Override
+  protected ServiceName getServiceName() {
+    return ServiceName.KEY_MANAGER;
+  }
+
+  @Override
+  protected String getFriendlyServiceName() {
+    return ServiceName.KEY_MANAGER_AUTH.toString();
+  }
+
+  @Override
+  protected String getMinVersion() {
+    if (currentVersion != null) {
+      return properties.getKeyManagerAuth().getMinVersion();
+    }
+
+    return null;
   }
 
   /**
@@ -67,7 +86,8 @@ public class KmAuthHealthIndicator extends AuthenticationServiceHealthIndicator 
    * @return POD aggregated JSON object or null if the application cannot retrieve POD HC
    */
   private JsonNode readAggregatedHC() {
-    String aggregatedResponse = getHealthResponse(properties.getSymphonyUrl() + HC_AGGREGATED_URL_PATH);
+    String aggregatedResponse =
+        getHealthResponse(properties.getSymphonyUrl() + HC_AGGREGATED_URL_PATH);
 
     if (aggregatedResponse != null) {
       return parseJsonResponse(aggregatedResponse);
@@ -77,9 +97,10 @@ public class KmAuthHealthIndicator extends AuthenticationServiceHealthIndicator 
   }
 
   /**
-   * Retrieve connectivity status according to the health-check JSON object
-   * @param jsonNode Health-check JSON object
-   * @return Connectivity status
+   * -   * Retrieve connectivity status according to the health-check JSON object
+   * -   * @param jsonNode Health-check JSON object
+   * -   * @return Connectivity status
+   * -
    */
   private Status retrieveConnectivityStatus(JsonNode jsonNode) {
     if (jsonNode == null) {
@@ -93,35 +114,6 @@ public class KmAuthHealthIndicator extends AuthenticationServiceHealthIndicator 
     }
 
     return Status.DOWN;
-  }
-
-  @Override
-  protected ServiceName getServiceName() {
-    return ServiceName.KEY_MANAGER;
-  }
-
-  @Override
-  protected String getFriendlyServiceName() {
-    return ServiceName.KEY_MANAGER_AUTH.toString();
-  }
-
-  @Override
-  protected String getMinVersion() {
-    if (currentVersion != null) {
-      return properties.getKeyManagerAuth().getMinVersion();
-    }
-
-    return null;
-  }
-
-  @Override
-  protected String getServiceBaseUrl() {
-    return properties.getKeyManagerAuthUrl();
-  }
-
-  @Override
-  protected String getServiceField() {
-    return SERVICE_FIELD;
   }
