+++ integration-healthcheck/src/test/java/org/symphonyoss/integration/healthcheck/services/indicators/KmHealthIndicatorTest.java
@@ -14,49 +14,41 @@
  * limitations under the License.
  */
 
-package org.symphonyoss.integration.healthcheck.services;
+package org.symphonyoss.integration.healthcheck.services.indicators;
 
 import static org.junit.Assert.assertEquals;
 
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.Mock;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringRunner;
-import org.springframework.test.util.ReflectionTestUtils;
 import org.symphonyoss.integration.authentication.AuthenticationProxy;
 import org.symphonyoss.integration.authentication.api.enums.ServiceName;
-import org.symphonyoss.integration.event.MessageMLVersionUpdatedEventData;
+import org.symphonyoss.integration.healthcheck.services.indicators.KmHealthIndicator;
 import org.symphonyoss.integration.logging.LogMessageSource;
-import org.symphonyoss.integration.model.message.MessageMLVersion;
 import org.symphonyoss.integration.model.yaml.IntegrationProperties;
 
 /**
- * Test class to validate {@link AgentHealthIndicator}
+ * Test class to validate {@link KmHealthIndicator}
  * Created by rsanchez on 23/11/16.
  */
 @RunWith(SpringRunner.class)
 @SpringBootTest
 @EnableConfigurationProperties
-@ContextConfiguration(classes = {IntegrationProperties.class, AgentHealthIndicator.class})
-public class AgentHealthIndicatorTest {
+@ContextConfiguration(classes = {IntegrationProperties.class, KmHealthIndicator.class})
+public class KmHealthIndicatorTest {
 
-  private static final String MOCK_VERSION = "1.45.0-SNAPSHOT";
+  private static final String MOCK_VERSION = "1.45.0";
 
-  private static final ServiceName SERVICE_NAME = ServiceName.AGENT;
+  private static final ServiceName SERVICE_NAME = ServiceName.KEY_MANAGER;
 
-  private static final String AGENT_MESSAGEML_VERSION2 = "1.46.0";
+  private static final String MOCK_SERVICE_URL = "https://nexus.symphony.com:443/relay";
 
-  private static final String AGENT_MESSAGEML_VERSION2_SNAPSHOT = "1.46.0-SNAPSHOT";
-
-  private static final String MOCK_SERVICE_URL = "https://nexus.symphony.com:8444/agent";
-
-  private static final String MOCK_HC_URL = MOCK_SERVICE_URL + "/v1/HealthCheck";
+  private static final String MOCK_HC_URL = MOCK_SERVICE_URL + "/HealthCheck/version";
 
   @MockBean
   private AuthenticationProxy authenticationProxy;
@@ -65,14 +57,7 @@ public class AgentHealthIndicatorTest {
   private LogMessageSource logMessageSource;
 
   @Autowired
-  private AgentHealthIndicator indicator;
-
-  private MockApplicationPublisher<MessageMLVersionUpdatedEventData> publisher = new MockApplicationPublisher<>();
-
-  @Before
-  public void init() {
-    ReflectionTestUtils.setField(indicator, "publisher", publisher);
-  }
+  private KmHealthIndicator indicator;
 
   @Test
   public void testHealthCheckUrl() {
@@ -93,5 +78,4 @@ public class AgentHealthIndicatorTest {
   public void testServiceBaseUrl() {
     assertEquals(MOCK_SERVICE_URL, indicator.getServiceBaseUrl());
   }
-
 }
